Fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,13 +6,13 @@
  */
 int main(void)
 {
-	int i = 1, u2, u1 = 2, u = 1;
+	long int i = 1, u2, u1 = 2, u = 1;
 
 	printf("%d, %d", u, u1);
 	while (i < 49)
 	{
 		u2 = u1 + u;
-		printf(", %d", u2);
+		printf(", %ld", u2);
 		u = u1;
 		u1 = u2;
 		i++;
